-
-Changed ServerMachineState.getState() to byte from int
-Removed error logging try/catch clauses from ActionServer.java
-Adjusted ActionServer.java to byte usage in ServerMachineState.getState()

+++ src/main/java/com/github/rosjava_actionlib/ServerStateMachine.java
@@ -63,7 +63,11 @@ final class ServerStateMachine {
         this.state = GoalStatus.PENDING;
     }
 
-    public final synchronized int getState() {
+    /**
+     *
+     * @return
+     */
+    public final synchronized byte getState() {
         return this.state;
     }
